refactor: auto-split test_mesh_generators.py
